refactor(grid): extract scroll input, focus tracking, and default constants
