+++ .github/copilot-instructions.md
@@ -42,7 +42,7 @@ This project's AI knowledge is organized in four tiers to minimize context windo
 > - **`DECIDED` is one bullet, not a paragraph.** Format: `DECIDED (date/PR): <conclusion>. WHY: <one-line rationale>. <File / test reference>.` Aim for 1-4 lines. If you have more to say, the next bullet is `INVARIANT:` or `TENSION:`, not a continuation paragraph.
 > - **No history-as-DECIDED.** "We tried X, then Y, then settled on Z because A then B then C" is a commit message, not knowledge. Write the conclusion ("Use Z; X breaks because A") and let `git log` carry the story. The exception is a "RULED OUT:" line listing alternatives that look obvious but don't work — one line each, max.
 > - **Preserve every spec file path, PR/issue number, function/class name, and "MUST/MUST NOT" rule.** Those are the navigation targets. Drop adjectives, transitions, and explanation of obvious mechanics — keep proper nouns.
-> - **Tables beat prose for enumerations.** A 23-row feature/binding list is a 25-line table, not a 25-line paragraph.
+> - **Tables beat prose for enumerations.** A 23-row feature/binding list is a 25-line table, not a 25-line paragraph. **But watch the byte cost:** Prettier re-pads every cell to the widest one, so a table with one long column (e.g. an event list) can cost 2× its content in whitespace. When one column is much wider than the rest, use a bullet list (`- name → a, b, c`) and state any uniform column as a single sentence above the list. Always re-run `bunx prettier --write` before measuring `wc -c`.
 > - **Compress, don't append.** When updating an existing entry, rewrite it. Do not stack a new "DECIDED (May 2026, follow-up to follow-up): ..." on top — fold the new fact into the existing entry.
 > - **No tier-by-tier history sections.** When a multi-pass refactor lands, replace the per-pass narrative with the post-refactor invariants. Keep at most one short paragraph of historical context if a future maintainer would otherwise be confused by the absence.
 > - **Move guidance to the right home.** Test-writing patterns belong in `.github/instructions/testing-patterns.instructions.md` (auto-applied to `*.spec.ts`), not in a knowledge file. Workflow recipes belong in `.github/skills/`. Knowledge files describe _how the code is and why_, not _how to write tests for it_.
