+++ .github/knowledge/grid-plugins.md
@@ -20,6 +20,8 @@ related: [grid-plugins-catalog-data, grid-plugins-catalog-ui, grid-plugins-shell
 - FLOW: `attach(grid)` → merge defaults + user config → store grid ref → `onPluginAttached()` notifications → [runtime hooks] → `detach()` → abort signal fires → cleanup.
 - INVARIANT: `disconnectSignal` (AbortSignal) fires on detach — use it for ALL event-listener cleanup.
 - INVARIANT: `plugin.grid` is available after `attach()`, null after `detach()`.
+- INVARIANT: `plugin.grid` (typed `GridElement`) is the object plugins read internal state from (`_sortState`, `_visibleColumns`, `_virtualization`, …). Do **NOT** re-route those reads through `this.gridElement` / `grid._hostElement`. Production satisfies `_hostElement === this` (`grid.ts` `get _hostElement() { return this; }`), but plugin specs mock `grid` as a plain object with **no** `_hostElement`, so the swap yields `undefined` at runtime. Verified: 73 failures across 7 specs (tree, undo-redo, reorder-columns, multi-sort, grouping-rows, row-drag-drop, server-side).
+- DECIDED (Aug 2026, PR #464 review follow-up): the ~13 `this.grid as unknown as GridHost` casts in plugins **stay**, each centralized in one `#internalGrid` accessor per plugin. WHY: `GridElement extends GridElementRef` (`core/plugin/types.ts`) and `GridHost = InternalGrid & HTMLElement` (`core/types.ts`) are parallel, unrelated descriptions of the same runtime object, so a single `as` fails comparability. RULED OUT: (a) casting from `gridElement` — breaks 73 tests, adds a runtime `_hostElement` coupling for zero functional gain; (b) `interface GridElement extends GridElementRef, InternalGrid` — same-named members conflict (`_visibleColumns: ColumnConfig[]` vs `ColumnInternal<T>[]`, `rows: unknown[]` vs `T[]`) and TS requires identical types; (c) hoisting one `protected get internalGrid()` onto `BaseGridPlugin` — the base-class member name is unmanglable at every call site, whereas per-plugin `#internalGrid` mangles to one char (editing has only ~500 B of budget headroom). Removing them properly requires reconciling `GridElementRef` vs `InternalGrid` — a type-architecture change, not a cleanup.
 - DECIDED: prefer config-driven init over post-ready imperative setup. Known-at-config-time resources accept a config prop and auto-init in `attach()` (pattern: `ServerSidePlugin` reads `config.dataSource`). Reserve imperative setters (`setDataSource()`) for runtime swaps.
 - DECIDED (Feb 2026, dead-config audit): a key declared on `*PluginConfig` + set in `defaultConfig` + documented, with ZERO consumers, is **dead and MUST be removed** — never "reserved for future use". Sweep with grep for `this.config.<key>`, `config.<key>`, `gridConfig.<key>` across the plugin dir + core. Removed: `visibility.allowHideAll`, `master-detail.collapseOnClickOutside`, `filtering.{trimInput,useWorker}`. PENDING: `pinned-rows.{showRowCount,showSelectedCount,showFilteredCount}` (removal needs a synthesized default `slots: []`).
