+++ libs/grid/src/public.ts
@@ -223,14 +223,10 @@ export type {
   UpdateSource,
 } from './lib/core/types';
 
-// Re-export FitModeEnum for runtime usage
-export {
-  ColumnInferenceModeEnum,
-  DEFAULT_A11Y_MESSAGES,
-  DEFAULT_ANIMATION_CONFIG,
-  DEFAULT_GRID_ICONS,
-  FitModeEnum,
-} from './lib/core/types';
+// Runtime (non-type) exports: enum-like consts stay in types.ts because their
+// `type` aliases derive from them via `typeof`; default values live in defaults.ts.
+export { DEFAULT_A11Y_MESSAGES, DEFAULT_ANIMATION_CONFIG, DEFAULT_GRID_ICONS } from './lib/core/defaults';
+export { ColumnInferenceModeEnum, FitModeEnum } from './lib/core/types';
 
 // Re-export sorting utilities for custom sort handlers
 export { builtInSort, defaultComparator } from './lib/core/internal/sorting';
